+++ scripts/template/package.json
@@ -1,12 +1,14 @@
 {
   "name": "nissix-__name__",
   "version": "1.0.0",
-  "description": "A plugin for Nissix called __name__",
-  "main": "dist/index.js",
-  "bin": "bin.js",
+  "description": "__name__ a plugin for Nissix ",
+  "bin": "",
   "author": "",
   "license": "ISC",
   "scripts": {
-    "bin": "echo \"starting __name__\" && exit 1"
+    "build": "",
+    "watch": "",
+    "clean": "rm -rf dist/",
+    "start": "echo \"starting __name__\" && exit 1"
   }
 }
